requirements.txt
# CTR-GCN AVA Action Recognition
# Core dependencies (pin exact versions for reproducibility)

# PyTorch 2.2 + CUDA 12.1 (install via --index-url)
# torch==2.2.1
# torchvision==0.17.1
# torchaudio==2.2.1

# OpenMMLab (install with CUDA wheel URL)
# mmengine==0.10.3
# mmcv==2.1.0
# mmdet==3.3.0
# mmpose==1.3.1

# Core science
numpy==1.26.4
pandas==2.2.0
scipy==1.12.0
scikit-learn==1.4.0
einops==0.7.0

# Video / download
yt-dlp==2024.3.10
ffmpeg-python==0.2.0
boto3==1.34.0
gdown==5.1.0
requests==2.31.0
opencv-python-headless==4.9.0.80

# Config + validation
pyyaml==6.0.1
jsonschema==4.21.0

# Logging + tracking
wandb==0.16.3
tensorboard==2.16.0
tqdm==4.66.2
rich==13.7.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0
pytest-timeout==2.2.0
